# src/Parbat.Core/Parbat.Core.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Parbat.Core.API",
    "description": "An open source ERP for Universities",
    "contact": {
      "name": "Dr. Muhammad Qasim Pasta",
      "url": "https://twitter.com/mqpasta",
      "email": "qasim.pasta@ieee.org"
    },
    "license": {
      "name": "The MIT License (MIT)",
      "url": "http://opensource.org/licenses/mit-license.php"
    },
    "version": "v2.0"
  },
  "paths": {
...
    "/api/v2.0/ModuleAssessment/Get/{id}": {
      "get": {
        "tags": [
          "ModuleAssessment"
        ],
        "summary": "Get a record",
        "operationId": "ModuleAssessmentGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleAssessment"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleAssessment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleAssessment"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.0/ModuleAssessment/Update": {
      "put": {
        "tags": [
          "ModuleAssessment"
        ],
        "summary": "Update a ModuleAssessment",
        "operationId": "ModuleAssessmentUpdate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleAssessment"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleAssessment"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleAssessment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleAssessment"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleAssessment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleAssessment"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/ModuleAssessment/Create": {
      "post": {
        "tags": [
          "ModuleAssessment"
        ],
        "summary": "Create",
        "operationId": "ModuleAssessmentCreate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleAssessment"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleAssessment"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleAssessment"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleAssessment"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleAssessment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleAssessment"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/ModuleAssessment/Delete/{id}": {
      "delete": {
        "tags": [
          "ModuleAssessment"
        ],
        "summary": "Delete",
        "operationId": "ModuleAssessmentDeleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v2.0/OrgUnitType/Get/{id}": {
      "get": {
        "tags": [
          "OrgUnitType"
        ],
        "summary": "Get a record",
        "operationId": "OrgUnitTypeGetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitType"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2.0/OrgUnitType/Update": {
      "put": {
        "tags": [
          "OrgUnitType"
        ],
        "summary": "Update a unittype",
        "operationId": "OrgUnitTypeUpdate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitType"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/OrgUnitType/Create": {
      "post": {
        "tags": [
          "OrgUnitType"
        ],
        "summary": "Create",
        "operationId": "OrgUnitTypeCreate",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitType"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgUnitType"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2.0/OrgUnitType/Delete/{id}": {
      "delete": {
        "tags": [
          "OrgUnitType"
        ],
        "summary": "Delete",
        "operationId": "OrgUnitTypeDeleteById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModuleAssessment": {
        "required": [
          "AssessmentCategoryID",
          "Name",
          "Weight"
        ],
        "type": "object",
        "properties": {
          "ModuleAssessmentID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "Name": {
            "minLength": 1,
            "type": "string"
          },
          "Weight": {
            "type": "number",
            "format": "double"
          },
          "AssessmentCategoryID": {
            "type": "integer",
            "format": "int64"
          },
          "Key": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "OrgUnitType": {
        "required": [
          "Name",
          "ShortName"
        ],
        "type": "object",
        "properties": {
          "OrgUnitTypeID": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "ShortName": {
            "minLength": 1,
            "type": "string"
          },
          "Name": {
            "minLength": 1,
            "type": "string"
          },
          "Key": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}